Get Posterior Predictive (PP)

This function takes the results from the brbr model and performs posterior predictive (PP) summaries for each Hydrological Unit Code (HUC12). The PP distribution is then split into three categories: Functioning Acceptably (FA), Functioning at Acceptable Risk (FAR), and Functioning at Unacceptable Risk (FUR) where the cutoffs are 0-0.1 (FA), 0.1-0.3 (FAR), and 0.3-1 (FAR).

Usage

btbr_pp(data, btbr_brm, indicator, ...)

Arguments

btbr_brm

A previously created brbr model.

indicator

A character defining the type of indicator, e.g. 'sediment', 'temperature'.

...

Arguments to pass to brms.

A

data.frame with appropriate covariates to predict with from previous model, e.g. natural_erosion, spec_delFS, ig_or_not, etc.

Value

A data.frame.

Note

The indicator argument is used to split the PP into custom thresholds previously defined.